Move `Buffer` into a separate module
